requirements_py310_mac_metal_gpu.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
db-dtypes==1.2.0
decorator==5.1.1
dill==0.3.8
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.4.0
gcsfs==2024.3.1
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.21.0
google-cloud-bigquery-storage==2.24.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.48.2
h5py==3.10.0
holidays==0.58
idna==3.6
jax==0.4.26
joblib==1.4.0
keras==2.13.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.7
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==21.3
pandarallel==1.6.5
pandas==1.5.3
proto-plus==1.23.0
protobuf==4.25.5
psutil==5.9.8
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydata-google-auth==1.8.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.13.0
tensorflow-metal==1.1.0
termcolor==2.4.0
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.5.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.14.1
yarl==1.9.4

# tensorflow-metal    # provides metal GPU support
